requirements.txt
certifi==2020.6.20
chardet==3.0.4
docutils==0.16
idna==2.10
Kivy==1.11.1
Kivy-Garden==0.1.4
Pygments==2.6.1
requests==2.24.0
urllib3==1.25.10
altgraph==0.17
ecdsa==0.15
esptool==2.8
macholib==1.14
pyaes==1.6.1
PyInstaller==3.6
pyserial==3.4
PySide2==5.15.0
PySimpleGUI==4.28.0
PySimpleGUIQt==0.35.0
shiboken2==5.15.0
six==1.15.0
